+++ tests/contained-value-set.spec.ts
@@ -155,7 +155,7 @@ test.describe(() => {
   test('should validate resource.id', async({page}) => {
     await PWUtils.getTableCell(flContainedTable, 2, 6).locator(editLoc).click();
     const dialog = page.locator('.cdk-dialog-container');
-    const idInput = dialog.getByLabel('Id', {exact: true});
+    const idInput = dialog.getByLabel('Id', { exact: true });
     await idInput.waitFor({ state: 'visible' }); // Wait until the input is visible
     await expect(idInput).toBeVisible();
     const idParent = idInput.locator('..');